feature: Proper (parallel!) decision loop, supporting SmartObjects feature: Basic SmartObjects skeleton feature: Half-proper event-based Action execution feature: u32 and Reflect support for ContextValues
I have realized that the only sensible implementation for a Utility Actions in an ECS library is via Events - we need Actions to have access to queries and whatnot.

This commit is the result of a long, long battle with the reflection system to make sure those events can both (a) be defined by the end-user & (b) reflected.

The current solution is crude; we provide 16 slots' worth (currently) of distinct but generic events within an enum. I have a better solution in mind, but this is TBD.