feature: Proper user Consideration support! refactor: Scoring phase replaced with a new Exclusive System-based approach refactor: Consideration requests are now batched per AI+Action
Considerations are now modelled as user Systems. User code is expected to register all Consideration systems as SystemIds to a new Resource.

The resource is then consumed by an Exclusive System to run the Systems in arbitrary order.

This is, tragically, a single-thread bottleneck right now, but uses all the Consideration optimization tricks to skip as much work as possible.