feature: LOD support for AIs + optimizations. fix: General warning/visibility cleanup. doc: Better docs for actions.rs and curves.rs
LODs are a Big Deal for running AIs at scale.

This is, for example, how an open-world game like Skyrim can have NPCs that truly travel around the world even when they are offscren.

It's also essential for 'dense' worlds with many SmartObjects available to each agent, like a whole city's worth in The Sims 3.